+++ Gameplan/UI/UiTextInput.h
@@ -5,7 +5,7 @@
 class UiTextInput : public UiInput
 {
 public:
-	UiTextInput(int _x, int _y, int _maxChar);
+	UiTextInput(float _x, float _y, unsigned int _maxChar);
 	virtual void draw(class Application* _application);
 	void setPlaceholder(std::string _placeholder);
 	std::string inputText = "";
@@ -13,7 +13,7 @@ public:
 
 private:
 	std::string placeholder = "";
-	int maxChar;
+	unsigned int maxChar;
 	int charAmount = 0;
 	float timePassedFromLastInput;
 	bool isFirstChar = true;
